# opt: optimize stats and FD computation