--- a/pkg/sql/opt/props/col_stats_map_test.go
+++ b/pkg/sql/opt/props/col_stats_map_test.go
@@ -57,9 +57,11 @@ func TestColStatsMap(t *testing.T) {
 		{cols: []opt.ColumnID{3, 4}, expected: "(5)+(1)+(1,5)+(2)+(1,2)+(3,4)"},
 	}
 
+	var maxColID opt.ColumnID
 	tcStats := make([]props.ColStatsMap, len(testcases))
 	// First calculate the stats for all steps, making copies every time. This
 	// also tests that the stats are copied correctly and there is no aliasing.
+	// Also find the maximum column ID used.
 	for tcIdx, tc := range testcases {
 		stats := &tcStats[tcIdx]
 		if tcIdx > 0 {
@@ -75,11 +77,17 @@ func TestColStatsMap(t *testing.T) {
 		} else {
 			stats.RemoveIntersecting(cols)
 		}
+		for _, col := range tc.cols {
+			if col > maxColID {
+				maxColID = col
+			}
+		}
 	}
 
 	for tcIdx, tc := range testcases {
 		stats := &tcStats[tcIdx]
 		var b strings.Builder
+		var singletonsInMap opt.ColSet
 		for i := 0; i < stats.Count(); i++ {
 			get := stats.Get(i)
 			if i != 0 {
@@ -92,7 +100,27 @@ func TestColStatsMap(t *testing.T) {
 				t.Errorf("could not find cols in map: %s", get.Cols)
 			}
 			if get != lookup {
-				t.Errorf("lookup did not return expected colstat: %+v vs. %+v", get, lookup)
+				t.Errorf("Lookup did not return expected colstat: %+v vs. %+v", get, lookup)
+			}
+			if get.Cols.Len() == 1 {
+				col := get.Cols.SingleColumn()
+				singletonsInMap.Add(col)
+				lookup, ok := stats.LookupSingleton(col)
+				if !ok {
+					t.Errorf("LookupSingleton(%d) could not find col in map", col)
+				}
+				if get != lookup {
+					t.Errorf("LookupSingleton did not return expected colstat: %+v vs. %+v", get, lookup)
+				}
+			}
+		}
+
+		// Lookup all possible singletons with LookupSingleton.
+		for col := opt.ColumnID(1); col <= maxColID; col++ {
+			_, ok := stats.LookupSingleton(col)
+			expected := singletonsInMap.Contains(col)
+			if expected != ok {
+				t.Errorf("expected LookupSingleton(%d) to return ok=%t, got ok=%t", col, expected, ok)
 			}
 		}