#### opt: make ColSet alias safer

This commit makes a minor change to the aliasing of an `opt.ColSet` that
makes it resilient to future misuse. A comment has been added for
additional clarity.

Release note: None

#### opt: fix bug in ColStatsMap.LookupSingleton

This commit fixes a bug in `(*ColStatsMap).LookupSingleton` that caused
"index out-of-bounds" errors when looking up a column that is not in the
map, but is a prefix of a column set in the map. For example, if the map
`m` contained the column set `(2,3)`, `m.LookupSingleton(2)` would err.

There is no release note because the bug is not present in any releases.

Release note: None

#### opt: add slow-query-7 benchmark

This commit add a `slow-query-7` benchmark which is based on
`slow-query-6` and contains a CTE.

Release note: None

#### opt: use ColStatsMap.LookupSingleton in more places

This commit replaces uses of `(*ColStatsMap).Lookup` with
`(*ColStatsMap).LookupSingleton` in a handful of places. This avoids
allocations for single-column sets of column IDs that exceed 128 (the
maximum column ID that fits in the stack-allocated part of an
`opt.ColSet`).

Release note: None

#### opt: optimize building WithScan functional dependencies

A WithScan expression's functional dependencies are inherited from the
underlying With expression, with the columns remapped. Prior to this
commit, the remapping was performed by first copying the With
expression's FDs, adding equivalencies between each corresponding column
in with With and WithScan, and finally projecting away the With's
columns. The latter two steps can be particularly expensive when FDs
contain column IDs greater than 128. Now, the more efficient
`(*FuncDeps).RemapFrom` method is used.

Epic: None

Release note: None