@mgartner 2023-12-21
Benchmark results are below. Notice the large benefit for slow-query-7. It looks like slow-query-2 got a bit slower and has additional allocations. I'm not exactly sure why at this point, but I'll try to dig in and understand.

```
name                                        old time/op    new time/op    delta
SlowQueries/slow-query-1/reorder-join-0-24    2.13ms ± 0%    2.10ms ± 0%   -1.76%  (p=0.008 n=5+5)
SlowQueries/slow-query-1/reorder-join-8-24    27.6ms ± 1%    27.2ms ± 1%   -1.26%  (p=0.032 n=5+5)
SlowQueries/slow-query-2/reorder-join-0-24    5.33ms ± 5%    5.66ms ± 1%   +6.20%  (p=0.008 n=5+5)
SlowQueries/slow-query-2/reorder-join-8-24     618ms ± 1%     612ms ± 1%   -1.11%  (p=0.032 n=5+5)
SlowQueries/slow-query-3/reorder-join-0-24    89.8ms ± 1%    87.3ms ± 0%   -2.79%  (p=0.016 n=5+4)
SlowQueries/slow-query-3/reorder-join-8-24     105ms ± 0%     102ms ± 1%   -2.68%  (p=0.008 n=5+5)
SlowQueries/slow-query-4/reorder-join-0-24    11.4ms ± 1%    11.3ms ± 1%     ~     (p=0.151 n=5+5)
SlowQueries/slow-query-4/reorder-join-8-24     604ms ± 1%     604ms ± 1%     ~     (p=1.000 n=5+5)
SlowQueries/slow-query-5/reorder-join-0-24    80.6ms ± 1%    79.7ms ± 1%   -1.19%  (p=0.008 n=5+5)
SlowQueries/slow-query-5/reorder-join-8-24     1.49s ± 1%     1.48s ± 1%     ~     (p=0.690 n=5+5)
SlowQueries/slow-query-6/reorder-join-0-24    52.5ms ± 0%    52.0ms ± 1%   -1.08%  (p=0.008 n=5+5)
SlowQueries/slow-query-6/reorder-join-8-24     2.10s ± 2%     2.10s ± 2%     ~     (p=0.841 n=5+5)
SlowQueries/slow-query-7/reorder-join-0-24     584ms ± 0%      73ms ± 1%  -87.41%  (p=0.008 n=5+5)
SlowQueries/slow-query-7/reorder-join-8-24     3.27s ± 1%     2.64s ± 1%  -19.16%  (p=0.008 n=5+5)

name                                        old alloc/op   new alloc/op   delta
SlowQueries/slow-query-1/reorder-join-0-24     680kB ± 0%     679kB ± 0%   -0.15%  (p=0.008 n=5+5)
SlowQueries/slow-query-1/reorder-join-8-24    6.90MB ± 0%    6.89MB ± 0%   -0.25%  (p=0.016 n=4+5)
SlowQueries/slow-query-2/reorder-join-0-24    1.14MB ± 0%    1.19MB ± 0%   +4.08%  (p=0.016 n=4+5)
SlowQueries/slow-query-2/reorder-join-8-24    74.4MB ± 0%    74.4MB ± 0%     ~     (p=0.548 n=5+5)
SlowQueries/slow-query-3/reorder-join-0-24    43.2MB ± 0%    42.2MB ± 0%   -2.35%  (p=0.008 n=5+5)
SlowQueries/slow-query-3/reorder-join-8-24    48.6MB ± 0%    47.6MB ± 0%   -2.15%  (p=0.008 n=5+5)
SlowQueries/slow-query-4/reorder-join-0-24    4.27MB ± 0%    4.27MB ± 0%   -0.04%  (p=0.008 n=5+5)
SlowQueries/slow-query-4/reorder-join-8-24     129MB ± 0%     129MB ± 0%   -0.02%  (p=0.008 n=5+5)
SlowQueries/slow-query-5/reorder-join-0-24    35.8MB ± 0%    35.8MB ± 0%   -0.09%  (p=0.008 n=5+5)
SlowQueries/slow-query-5/reorder-join-8-24     279MB ± 0%     278MB ± 0%   -0.09%  (p=0.008 n=5+5)
SlowQueries/slow-query-6/reorder-join-0-24    13.5MB ± 0%    13.3MB ± 0%   -0.95%  (p=0.016 n=4+5)
SlowQueries/slow-query-6/reorder-join-8-24     489MB ± 0%     487MB ± 0%   -0.40%  (p=0.008 n=5+5)
SlowQueries/slow-query-7/reorder-join-0-24    33.6MB ± 0%    18.3MB ± 0%  -45.63%  (p=0.016 n=5+4)
SlowQueries/slow-query-7/reorder-join-8-24     622MB ± 0%     611MB ± 0%   -1.80%  (p=0.008 n=5+5)

name                                        old allocs/op  new allocs/op  delta
SlowQueries/slow-query-1/reorder-join-0-24     3.97k ± 0%     3.95k ± 0%   -0.40%  (p=0.008 n=5+5)
SlowQueries/slow-query-1/reorder-join-8-24     70.1k ± 0%     69.9k ± 0%   -0.30%  (p=0.016 n=4+5)
SlowQueries/slow-query-2/reorder-join-0-24     5.43k ± 0%     5.78k ± 0%   +6.51%  (p=0.016 n=4+5)
SlowQueries/slow-query-2/reorder-join-8-24      592k ± 0%      592k ± 0%     ~     (p=0.548 n=5+5)
SlowQueries/slow-query-3/reorder-join-0-24      349k ± 0%      317k ± 0%   -9.08%  (p=0.008 n=5+5)
SlowQueries/slow-query-3/reorder-join-8-24      399k ± 0%      366k ± 0%   -8.20%  (p=0.008 n=5+5)
SlowQueries/slow-query-4/reorder-join-0-24     43.5k ± 0%     43.4k ± 0%   -0.11%  (p=0.008 n=5+5)
SlowQueries/slow-query-4/reorder-join-8-24     2.78M ± 0%     2.78M ± 0%   -0.03%  (p=0.008 n=5+5)
SlowQueries/slow-query-5/reorder-join-0-24      221k ± 0%      220k ± 0%   -0.43%  (p=0.008 n=5+5)
SlowQueries/slow-query-5/reorder-join-8-24     4.64M ± 0%     4.63M ± 0%   -0.16%  (p=0.008 n=5+5)
SlowQueries/slow-query-6/reorder-join-0-24      156k ± 0%      153k ± 0%   -2.01%  (p=0.016 n=4+5)
SlowQueries/slow-query-6/reorder-join-8-24     8.78M ± 0%     8.72M ± 0%   -0.69%  (p=0.008 n=5+5)
SlowQueries/slow-query-7/reorder-join-0-24      363k ± 0%      254k ± 0%  -30.07%  (p=0.016 n=5+4)
SlowQueries/slow-query-7/reorder-join-8-24     12.6M ± 0%     12.5M ± 0%   -0.74%  (p=0.008 n=5+5)
```

```
name                                                    old time/op    new time/op    delta
Phases/kv-read/Simple/Explore-24                          50.4µs ± 1%    50.3µs ± 0%    ~     (p=0.841 n=5+5)
Phases/kv-read-const/Simple/Explore-24                    50.8µs ± 1%    50.7µs ± 1%    ~     (p=1.000 n=5+5)
Phases/tpcc-new-order/Simple/Explore-24                    135µs ± 1%     132µs ± 1%  -1.95%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                     118µs ± 1%     116µs ± 1%  -1.32%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                  43.4µs ± 1%    43.0µs ± 1%    ~     (p=0.151 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                  603µs ± 0%     588µs ± 0%  -2.39%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24                258µs ± 1%     256µs ± 1%    ~     (p=0.151 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24        466µs ± 1%     465µs ± 0%    ~     (p=0.222 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24      412µs ± 0%     413µs ± 1%    ~     (p=0.841 n=5+5)
Phases/many-columns-and-indexes-b/Simple/Explore-24        557µs ± 0%     554µs ± 0%  -0.61%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24      469µs ± 0%     468µs ± 1%    ~     (p=0.095 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        289ms ± 2%     273ms ± 2%  -5.35%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      288ms ± 1%     278ms ± 1%  -3.62%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24        139µs ± 1%     137µs ± 1%  -1.41%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-24     58.8µs ± 1%    58.3µs ± 1%    ~     (p=0.095 n=5+5)
Phases/batch-insert-one/Simple/Explore-24                  116µs ± 1%     114µs ± 1%  -1.57%  (p=0.008 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                8.18ms ± 1%    8.02ms ± 0%  -1.94%  (p=0.008 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                   3.64ms ± 0%    3.61ms ± 0%  -0.76%  (p=0.008 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24      3.45ms ± 1%    3.41ms ± 0%  -0.93%  (p=0.008 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24    1.50ms ± 1%    1.48ms ± 1%  -1.34%  (p=0.032 n=5+5)

name                                                    old alloc/op   new alloc/op   delta
Phases/kv-read/Simple/Explore-24                          15.3kB ± 0%    15.3kB ± 0%    ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                    15.4kB ± 0%    15.4kB ± 0%    ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                   32.7kB ± 0%    32.7kB ± 0%    ~     (p=0.087 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                    27.0kB ± 0%    27.0kB ± 0%    ~     (p=0.063 n=5+4)
Phases/tpcc-delivery/Prepared/Explore-24                  12.8kB ± 0%    12.8kB ± 0%    ~     (p=0.429 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                  130kB ± 0%     130kB ± 0%    ~     (p=0.690 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24               58.9kB ± 0%    58.9kB ± 0%    ~     (p=0.333 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24       39.6kB ± 0%    39.6kB ± 0%    ~     (p=0.968 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24     5.50kB ± 0%    5.50kB ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24       47.9kB ± 0%    47.9kB ± 0%    ~     (p=0.865 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24     8.02kB ± 0%    8.02kB ± 0%    ~     (p=1.000 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        105MB ± 0%     105MB ± 0%  -0.06%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      104MB ± 0%     104MB ± 0%  -0.05%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24       29.7kB ± 0%    29.7kB ± 0%    ~     (p=0.333 n=5+4)
Phases/single-col-histogram-range/Prepared/Explore-24     13.0kB ± 0%    13.0kB ± 0%    ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                 64.9kB ± 0%    64.9kB ± 0%    ~     (p=0.754 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                3.94MB ± 0%    3.94MB ± 0%    ~     (p=0.421 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                    945kB ± 0%     945kB ± 0%    ~     (p=0.683 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24       926kB ± 0%     926kB ± 0%    ~     (p=0.222 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     736kB ± 0%     736kB ± 0%    ~     (p=0.310 n=5+5)

name                                                    old allocs/op  new allocs/op  delta
Phases/kv-read/Simple/Explore-24                            95.0 ± 0%      95.0 ± 0%    ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                      95.0 ± 0%      95.0 ± 0%    ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                      175 ± 0%       175 ± 0%    ~     (all equal)
Phases/tpcc-delivery/Simple/Explore-24                       179 ± 0%       179 ± 0%    ~     (all equal)
Phases/tpcc-delivery/Prepared/Explore-24                    75.0 ± 0%      75.0 ± 0%    ~     (all equal)
Phases/tpcc-stock-level/Simple/Explore-24                    707 ± 0%       707 ± 0%    ~     (all equal)
Phases/tpcc-stock-level/Prepared/Explore-24                  341 ± 0%       341 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-a/Simple/Explore-24         81.0 ± 0%      81.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/Explore-24       31.0 ± 0%      31.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24          136 ± 0%       136 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Prepared/Explore-24       49.0 ± 0%      49.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-c/Simple/Explore-24         683k ± 0%      681k ± 0%  -0.26%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24       678k ± 0%      676k ± 0%  -0.26%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24          522 ± 0%       522 ± 0%    ~     (all equal)
Phases/single-col-histogram-range/Prepared/Explore-24        251 ± 0%       251 ± 0%    ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                    331 ± 0%       331 ± 0%    ~     (all equal)
Phases/batch-insert-many/Simple/Explore-24                 29.5k ± 0%     29.5k ± 0%    ~     (p=1.000 n=5+4)
Phases/ored-preds-100/Simple/Explore-24                    4.53k ± 0%     4.53k ± 0%    ~     (p=1.000 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24       4.53k ± 0%     4.53k ± 0%    ~     (p=0.167 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     2.88k ± 0%     2.88k ± 0%    ~     (all equal)
```